Don't believe frame geometry less than 2x2
This looks very strange, but it appears that Qt can also give frameWidth=1 or
frameHeight=1 (usually simultaneously). This doesn't work well for Oxygen, at
least for height. Nor does it actually make any sense in most cases, since
window decorations are much higher than 1 pixel (for decorated windows). So the
ugly fix has to be updated to also check for width or height equal to 1, not
only 0.
This problem might actually be a bug in KWin from KF5, which seems to return
bogus values for _NET_FRAME_EXTENTS sometimes. If I switch to e.g. xfwm4, the
problem doesn't reproduce. Nevertheless, until it's proven so, let this
workaround be here.