Switch from QWeakPointer to QPointer
QWeakPointer::data() and QWeakPointer::QWeakPointer(QObject*) are
extensively used here, but both are deprecated in Qt5 and removed in
Qt6. QPointer retains this functionality.

Still, we have to use QWeakPointer<const XYZ> instead of
QPointer<const XYZ> on Qt4, because the latter doesn't compile.

Also, on Qt4 there's no QPointer::clear, so we work around it by
assigning default-constructed QPointer to the existing QPointer.